Update poiseuille.cpp
There are to adapatation, with or without hanging nodes. Still the  part of hanging nodes is not working. To decide between one or another, use the bool variable hanging_nodes. 
On the other hand, gamma error is not converging, so I need to come bcak to the old code to see the difference, since it used to converge ( without adaptation).